Fix type mismatch

+++ IndieGameZone.Infrastructure/Repositories/TransactionRepository.cs
@@ -37,7 +37,7 @@ namespace IndieGameZone.Infrastructure.Repositories
 			return PagedList<Transactions>.ToPagedList(transactionEntities, transactionParameters.PageNumber, transactionParameters.PageSize, ct);
 		}
 
-        public async Task<bool> IsOrderCodeExistsAsync(int orderCode, CancellationToken ct = default)
+        public async Task<bool> IsOrderCodeExistsAsync(long orderCode, CancellationToken ct = default)
         {
             return await AppDbContext.Transactions.AnyAsync(t => t.OrderCode == orderCode, ct);
         }
